requirements3.txt
requests
impacket
slackclient
progressbar
numpy
pillow
pytube3
PyCrypto
qrcode
base58
librosa
luhn
names
Faker
